pyproject.toml
[tool.poetry]
name = "urfu-schedule-bot"
version = "0.1.0"
description = "Telegram bot for UrFU schedule"
authors = ["Ваше Имя <email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

# Зависимости
[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^22.5"  # Обратите внимание на версию! v20.x — это последняя основная ветка с асинхронным API по умолчанию.
python-dotenv = "1.2.1"
loguru = "^0.7.3"
httpx = "^0.28.1"

# Группа зависимостей для разработки
vk-api = "^11.10.0"
[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^0.26.0"
pytest-cov = "^4.1.0"
black = "^24.10.0"
isort = "^5.13.2"
taskipy = "1.14.1"

# Скрипты
[tool.poetry.scripts]
bot = "src.bot:main"
bot_vk = "src.botvk:main"
test_module = "src.test_module:main"

[tool.taskipy.tasks]
# Тестирование
test = "pytest"
test-cov = "pytest --cov=src --cov-report=term-missing"
test-fast = "pytest -xvs"  # Быстрые тесты с детальным выводом
test-all = "pytest --cov=src --cov-report=html --cov-report=term-missing"

# Форматирование
format = "black src tests && isort src tests"
format-check = "black --check src tests && isort --check-only src tests"
lint = "flake8 src tests"
type-check = "mypy src"  # Если используете типизацию

# Качество кода
qa = "black --check src tests && isort --check-only src tests && flake8 src tests && pytest"


# Конфигурация pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-v",
    "--tb=short"
]

# Конфигурация black
[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# Конфигурация isort
[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["src"]

# Конфигурация mypy
[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"